TWINS-282: Refactor storage abstractions into direct implementations
Removed `StoragerAbstractLocal` and `StoragerAbstractS3` classes, integrating their methods directly into `StoragerLocal` and `StoragerS3`. This simplifies the hierarchy, making the implementations more cohesive and reducing unnecessary abstraction layers.
